agent: Add unit tests for `namespace.rs`
This patch adds a unit test for `namespace.rs`. The baseline for this
test comes from the Go agent tests. It turns out that the Go version of
this test creates a temp folder for `ipc` and `uts`, however these folders
doesn't has an effect in the tested code, therefore the implementation of
such temporal folders was removed.

Fixes: #50

Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>